WIP: Create consensus-encoding crate
Just a quick demo of what creating the `bitcoin-consensus-encoding`
crate will involve.

I'm hitting orphin rule violations that I can't debug right at this
minute, compiler is complaining about a `base58::Vec` type, no clue what
that is.
