+++ consensus-encoding/README.md
@@ -1,0 +1,16 @@
+Bitcoin Consensus Encoding
+==========================
+
+This crate provides encoding and decoding of types according to Bitcoin consensus rules. Crate
+works with`no_std` but requires an allocator.
+
+
+## Minimum Supported Rust Version (MSRV)
+
+This library should always compile with any combination of features on **Rust 1.56.1**.
+
+
+## Licensing
+
+The code in this project is licensed under the [Creative Commons CC0 1.0 Universal license](LICENSE).
+We use the [SPDX license list](https://spdx.org/licenses/) and [SPDX IDs](https://spdx.dev/ids/).
