+++ consensus-encoding/contrib/test_vars.sh
@@ -1,0 +1,10 @@
+#!/usr/bin/env bash
+
+# Test all these features with "std" enabled.
+FEATURES_WITH_STD="serde"
+
+# Test all these features without "std" enabled.
+FEATURES_WITHOUT_STD="alloc serde"
+
+# Run and lint these examples.
+EXAMPLES=""
